backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose[pycryptodome]
bcrypt
pycryptodome
cryptography
python-multipart
python-dotenv
